프론트 백 연결 완

+++ 백엔드/IntelliJ_ISOUR/IntelliJ/src/main/java/com/ISOUR/vo/MemberVO.java
@@ -6,7 +6,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class MemberVO {
-    private int id_num;
+    private Long id_num;
     private String id;
     private String pwd;
     private String name;
